added ThresholdBar.css, ThresholdBar.js without functionality
